fix an issue where properties that have been removed from the session are not removed in the DB
Due to the use of conflict: 'update' in the set method properties that have been removed from the session object won't be removed from the database. Setting this to conflict: 'replace' will resolve the issue and is safe due to the nature of express-session which pushes the new session object in it's entirety to the Store to be processed.